update emojis file structure & styling
